Menu 1
this menu allows for the interface to display Header, Header Line and Highlight bar with inverted color text.

This is the main code that started this battle. I will try and upload a short vidoe display the outcome using the adafruit ssd1306 and gfx library. The problem seems to happen with both Lines and the Highlight Bar... removing them and only using an arrow allows for working fuctionality so theirfor the problem lyes in the fill rectangle and draw line of either the gfx or adafruit.. this occurs on multiple ssd1306 so "cheapness isnt the issues expecialy due to the fact ive mixed and matched different panels with controller board and receiving different outcome. something has to be the case.. buffer? contrast? something is wrong making it not compatable with ALL SSD1306 OLEDs